requirements.txt
anyio==4.9.0
APScheduler==3.6.3
beautifulsoup4==4.13.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
ephem==4.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.3
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tornado==6.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
